% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/co2_growth_function.R
\name{co2_growth_function}
\alias{co2_growth_function}
\title{CO2 change mean \% since 1997}
\usage{
co2_growth_function(countryname)
}
\arguments{
\item{countryname}{}

\item{dataf}{}
}
\value{

}
\description{
CO2 change mean \% since 1997
}
