% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/co2_change-package.R
\docType{package}
\name{co2change-package}
\alias{co2change}
\alias{co2change-package}
\title{co2change: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
