% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_model_gas.R
\name{function_model_gas}
\alias{function_model_gas}
\title{Title}
\usage{
function_model_gas(countryname, gases)
}
\arguments{
\item{countryname}{}

\item{gases}{}

\item{datafr}{}
}
\value{

}
\description{
Title
}
