% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percentage_variation_decade.R
\name{percentage_variation_decade}
\alias{percentage_variation_decade}
\title{Title}
\usage{
percentage_variation_decade(decada, pais)
}
\arguments{
\item{decada}{}

\item{pais}{}
}
\value{

}
\description{
Title
}
