% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediccion_modelo_lineal.R
\name{prediccion_modelo_lineal}
\alias{prediccion_modelo_lineal}
\title{Prediction for decades 90', 00', 10' for greenhouse gases}
\usage{
prediccion_modelo_lineal(pais)
}
\arguments{
\item{pais}{}
}
\value{

}
\description{
Prediction for decades 90', 00', 10' for greenhouse gases
}
